feat(ipc): oauth2 token provider in AppState + fetch/invalidate commands + bindings
Adds Oauth2TokenProvider field to AppState (default + load), implements
auth_resolve_impl/auth_oauth2_fetch_token_impl/auth_invalidate_impl routing
OAuth2 through the provider, registers auth_oauth2_fetch_token + auth_invalidate
Tauri commands, fixes EnvVar struct literals missing `environments` field in
collection/env test helpers, and regenerates src/ipc/bindings.ts. OAuth2TokenInfoIpc
uses u32 (not u64) to satisfy specta's BigInt restriction.

+++ src-tauri/src/ipc/auth.rs
@@ -19,9 +19,10 @@ impl AuthCredentialsIpc {
 
 /// Result of a forced token fetch (the "Get token" button) — lifetime only;
 /// the token itself stays in the backend cache.
+/// `u32` (not u64) because specta forbids BigInt in generated TypeScript.
 #[derive(Debug, Clone, Serialize, Deserialize, Type)]
 pub struct OAuth2TokenInfoIpc {
-    pub expires_in_secs: u64,
+    pub expires_in_secs: u32,
 }
 
 #[cfg(test)]
